% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_hc.R
\name{additive_hc}
\alias{additive_hc}
\title{Calculates the additive proportion of the community effected}
\usage{
additive_hc(x, proportion = c(0.01, 0.05, 0.1, 0.2), resolution = 100)
}
\arguments{
\item{x}{A named list of ssd's of class "fitdists" as fitted by
\code{\link{ssd_fit_dists}} via package \code{\link{ssdtools}}.}

\item{proportion}{A numeric vector of proportion values to estimate hazard
concentrations for.}

\item{resolution}{The number of values over which to find estimates
large values will make the estimate more precise but will also be slow to run.}
}
\description{
This function calculate the additive proportion of the community effect
across two or more input ssds, as input via a named list.
}
