% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_hp.R
\name{additive_hp}
\alias{additive_hp}
\title{Calculates the additive proportion of the community effected}
\usage{
additive_hp(x, fixed_conc, ci = FALSE, ...)
}
\arguments{
\item{x}{A named list of ssd's of class "fitdists" as fitted by
\code{\link{ssd_fit_dists}} via package \code{\link{ssdtools}}.}

\item{fixed_conc}{List of concentrations series to use for additive HC
estimation}
}
\description{
This function calculate the additive proportion of the community effect
across two or more input ssds, as input via a named list.
}
