% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_predict.R
\name{additive_predict}
\alias{additive_predict}
\title{Calculates the additive proportion of the community effected}
\usage{
additive_predict(x, proportion = seq(0.01, 1 - 0.01, length = 100))
}
\arguments{
\item{x}{A named list of ssd's of class "fitdists" as fitted by
\code{\link{ssd_fit_dists}} via package \code{\link{ssdtools}}.}

\item{proportion}{A numeric vector of proportion values to estimate hazard
concentrations for.}
}
\description{
This function calculate the additive proportion of the community effect
across two or more input ssds, as input via a named list.
}
