% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggre_HU_Calc2.R
\name{aggre_HU_Calc2}
\alias{aggre_HU_Calc2}
\title{Title}
\usage{
aggre_HU_Calc2(
  CalcedHU,
  ChemData = Gross,
  zeros_PAF,
  aggrFUN = max,
  agg_jaar = F,
  TooLowLimit = 1e-04,
  National,
  verbose = FALSE
)
}
\description{
Title
}
