% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_limietsymbool}
\alias{filter_limietsymbool}
\title{Remove rows with Limietsymbool < or > and warn if any removed}
\usage{
filter_limietsymbool(
  inputData,
  inputwarnings,
  National,
  verbose,
  UsedParameters,
  OptionalParameters
)
}
\arguments{
\item{inputData}{The data.frame to check/modify}

\item{inputwarnings}{An InputWarnings R6 object}

\item{National}{Language string ("Nederlands" or "English")}

\item{verbose}{Logical, whether to add warnings}

\item{UsedParameters}{Character vector of required columns (for subsetting)}

\item{OptionalParameters}{Character vector of optional columns (for subsetting)}
}
\value{
The modified inputData (with limit rows removed)
}
\description{
Remove rows with Limietsymbool < or > and warn if any removed
}
