% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version_string.R
\name{get_package_version_string}
\alias{get_package_version_string}
\title{Get a Pretty Version String for an Installed Package}
\usage{
get_package_version_string(pkgname, National = "Nederlands")
}
\arguments{
\item{pkgname}{Character. Name of the installed package.}
}
\value{
Character string with the format \code{"pkgname version x.y.z (Git commit abcdefg)"}
  if the package was installed from GitHub or GitLab (and the commit hash is available),
  or \code{"pkgname version x.y.z"} otherwise. If the package is not installed, returns a message.
}
\description{
Returns a human-readable version string for an installed package, including the Git commit hash
if the package was installed from GitLab or GitHub.
}
\examples{
get_package_version_string("dplyr")
get_package_version_string("yourGitLabPackage")

}
