% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_grootheden_peculiarities}
\alias{handle_grootheden_peculiarities}
\title{Handle special grootheden and parameter code conversions}
\usage{
handle_grootheden_peculiarities(inputData, inputwarnings, National, verbose)
}
\arguments{
\item{inputData}{The data.frame containing the measurement data.}

\item{inputwarnings}{An InputWarnings R6 object for collecting warnings.}

\item{National}{The language string ("Nederlands" or "English").}

\item{verbose}{Logical; if TRUE, adds extra warnings for Corg and OS conversions.}
}
\value{
The modified inputData data.frame with the above conversions applied.
}
\description{
This function processes and corrects known peculiarities in the input data
for water quality measurements, specifically for certain grootheden/parameter
codes that require conversion or special handling according to the IM standard.
}
\details{
The following conversions are performed:
\describe{
  \item{T}{If Grootheid.code contains \code{"T"} but not \code{"Tw"}, all \code{"T"} are interpreted as water temperature (\code{"Tw"}).}
  \item{Corg}{If Parameter.code is \code{"Corg"} and Hoedanigheid.code ends with \code{"nf"}, these are recoded to \code{"DOC"} (Dissolved Organic Carbon) after filtration.}
  \item{OS}{If Parameter.code is \code{"OS"}, these are recoded to \code{"TSS"} (Total Suspended Solids).}
}

For each conversion, a warning is added to the inputwarnings object (in Dutch or English, as appropriate).
}
