% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_fun.R
\name{hc_fun}
\alias{hc_fun}
\title{Extracts the hazard concentration given the proportion effected.}
\usage{
hc_fun(x, proportion, ...)
}
\arguments{
\item{x}{A named list of ssd's of class "fitdists" as fitted by
\code{\link{ssd_fit_dists}} via package \code{\link{ssdtools}}. Alternatively,
a named list of concentration-response fits of class "bnecfit" as fitted by
\code{\link{bnec}} via package \code{\link{bayesnec}}.}

\item{proportion}{A numeric vector of proportion values to estimate hazard
concentrations for. These are proportion of the community in the case of a
fitted ssd, or effect proportions relative to the control in the case of a
concentration-response fitted model.}
}
\description{
Extracts the hazard concentration given the proportion effected.
}
\details{
This function calculates the hazard concentration, given a desired
proportion of the community effected or individual species endpoint effect.
}
