% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HU2msPAF.R
\name{HU2msPAF}
\alias{HU2msPAF}
\title{HU2msPAF}
\usage{
HU2msPAF(HU, TMOAname = "substance_key", groupName = "groep.fotoNL")
}
\arguments{
\item{HU}{data.frame with HU with at least the columns CAS/AquoCode, SampleID, ...}

\item{TMOAname}{name of column for toxic mode of action = "PrimaryMoA", NO
switched to AquoCode to force Response Addition}
}
\value{
data.frame with msPAF values
}
\description{
process Hazard Units (HU) for HU_Calc2() to msPAF values for each Sample; each UseClass ..?
}
\author{
jaap slootweg
}
