% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IMColNameMapper}
\alias{IMColNameMapper}
\title{IM Column Name Mapper (English <-> Dutch)}
\description{
IM Column Name Mapper (English <-> Dutch)

IM Column Name Mapper (English <-> Dutch)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IMColNameMapper-new}{\code{IMColNameMapper$new()}}
\item \href{#method-IMColNameMapper-to_dutch}{\code{IMColNameMapper$to_dutch()}}
\item \href{#method-IMColNameMapper-to_english}{\code{IMColNameMapper$to_english()}}
\item \href{#method-IMColNameMapper-clone}{\code{IMColNameMapper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IMColNameMapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-IMColNameMapper-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IMColNameMapper$new(mapping)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IMColNameMapper-to_dutch"></a>}}
\if{latex}{\out{\hypertarget{method-IMColNameMapper-to_dutch}{}}}
\subsection{Method \code{to_dutch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IMColNameMapper$to_dutch(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IMColNameMapper-to_english"></a>}}
\if{latex}{\out{\hypertarget{method-IMColNameMapper-to_english}{}}}
\subsection{Method \code{to_english()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IMColNameMapper$to_english(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IMColNameMapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IMColNameMapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IMColNameMapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
