% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputWarnings.R
\name{InputWarnings}
\alias{InputWarnings}
\title{Collect warnings to the inputwarnings data.frame}
\description{
Collect warnings to the inputwarnings data.frame

Collect warnings to the inputwarnings data.frame
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputWarnings-new}{\code{InputWarnings$new()}}
\item \href{#method-InputWarnings-add}{\code{InputWarnings$add()}}
\item \href{#method-InputWarnings-clone}{\code{InputWarnings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputWarnings-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputWarnings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputWarnings$new(init_df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputWarnings-add"></a>}}
\if{latex}{\out{\hypertarget{method-InputWarnings-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputWarnings$add(code, nl_text, en_text, National)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputWarnings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputWarnings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputWarnings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
