% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_percentile.R
\name{make_quantile}
\alias{make_quantile}
\title{Factory for Quantile Functions (with 'higher' method)}
\usage{
make_quantile(props)
}
\arguments{
\item{props}{Numeric vector of probabilities (between 0 and 1) at which to compute quantiles.}
}
\value{
A function that takes a numeric vector and returns the quantiles at the specified probabilities.
}
\description{
Creates a function that calculates specified quantiles (using the 'higher' method, type = 2) for a numeric vector.
}
\examples{
# Create a quantile function for 40\% and 80\%
q_fn <- make_quantile(c(0.4, 0.8))
# Use it on a numeric vector
x <- c(3, 7, 8, 12, 20)
q_fn(x)
}
