% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFase2.R
\name{CleanFase2}
\alias{CleanFase2}
\title{Clean and Prepare Input Data for SSD Analysis}
\usage{
CleanFase2(inputData, SSDsubstanceData, inputwarnings)
}
\arguments{
\item{inputData}{A `data.frame` containing measurement data. Must include columns such as \code{Concentration}, \code{substance_key}, \code{SampleID}, and \code{PreTreatment}.}

\item{SSDsubstanceData}{A `data.frame` with SSD substance information required for validation and processing.}
}
\value{
A processed \code{data.frame} with cleaned and aggregated measurement data, ready for SSD analysis.
}
\description{
This function cleans and processes the input data for SSD (Species Sensitivity Distribution) analysis.
It handles unit conversion, substance code checks, aggregation, and exclusion of rows with unknown SSD.
}
\details{
The function performs the following steps:
\itemize{
  \item Renames the \code{Concentration} column to \code{MeasuredValue}.
  \item Converts units and calculates concentrations using \code{convert_units_and_calculate_concentration}.
  \item Excludes rows with unknown or invalid SSD information.
  \item Checks and processes substance codes.
  \item Aggregates samples by taking the maximum value for each combination of \code{substance_key}, \code{SampleID}, and \code{PreTreatment}.
  \item Filters aggregated samples based on specific criteria for \code{PreTreatment}.
}
}
\examples{
\dontrun{
cleaned_data <- CleanFase2(inputData = my_data, SSDsubstanceData = ssd_data)
}

}
